More changes to events.

+++ Source/Client/Source/Modules/Objects/C_Maps.vb
@@ -400,7 +400,7 @@ Module C_Maps
                                             ReDim Map.Events(i).Pages(x).CommandList(y).Commands(Map.Events(i).Pages(x).CommandList(y).CommandCount)
                                             For z = 0 To Map.Events(i).Pages(x).CommandList(y).CommandCount
                                                 With Map.Events(i).Pages(x).CommandList(y).Commands(z)
-                                                    .Index = buffer.ReadInt32
+                                                    .Index = buffer.ReadByte
                                                     .Text1 = Trim(buffer.ReadString)
                                                     .Text2 = Trim(buffer.ReadString)
                                                     .Text3 = Trim(buffer.ReadString)
@@ -724,7 +724,7 @@ Module C_Maps
                                 If Map.Events(i).Pages(x).CommandList(y).CommandCount > 0 Then
                                     For z = 0 To Map.Events(i).Pages(x).CommandList(y).CommandCount
                                         With Map.Events(i).Pages(x).CommandList(y).Commands(z)
-                                            buffer.WriteInt32(.Index)
+                                            buffer.WriteByte(.Index)
                                             buffer.WriteString((.Text1))
                                             buffer.WriteString((.Text2))
                                             buffer.WriteString((.Text3))
